+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [v22.3.2] 2026-07-09
+
+- **CSD-121**
+  - Updates the CMR.js CMR client object to *not* store the CMR provider and update all class methods to require
+    a provider method parameter.    Update all tasks/related callers to provide expected input
+
 ## [v22.3.1] 2026-07-07
 
 - Bumped `archiver` version and corresponding updates to `bin/zip.js` to address CVE-2026-13149.
@@ -9985,8 +9991,9 @@ Note: There was an issue publishing 1.12.0. Upgrade to 1.12.1.
 
 ## [v1.0.0] - 2018-02-23
 
-[Unreleased]: https://github.com/nasa/cumulus/compare/v22.3.1...HEAD
-[v22.3.0]: https://github.com/nasa/cumulus/compare/v22.3.0...v22.3.1
+[Unreleased]: https://github.com/nasa/cumulus/compare/v22.3.2...HEAD
+[v22.3.2]: https://github.com/nasa/cumulus/compare/v22.3.1...v22.3.2
+[v22.3.1]: https://github.com/nasa/cumulus/compare/v22.3.0...v22.3.1
 [v22.3.0]: https://github.com/nasa/cumulus/compare/v22.2.5...v22.3.0
 [v22.2.5]: https://github.com/nasa/cumulus/compare/v22.2.4...v22.2.5
 [v22.2.4]: https://github.com/nasa/cumulus/compare/v22.2.3...v22.2.4